requirements.txt
# Dependencias de producción del servicio Cloud Run.
# Versiones fijadas para garantizar reproducibilidad en el build de Docker.
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-cloud-pubsub==2.27.0
google-cloud-bigquery==3.27.0
